remove extra files from wheel

+++ cmake/cmake_extension.py
@@ -215,8 +215,12 @@ class BuildExtension(build_ext):
         shutil.rmtree(f"{install_dir}/bin")
         shutil.rmtree(f"{install_dir}/share")
         shutil.rmtree(f"{install_dir}/lib/pkgconfig")
-        os.remove(f"{install_dir}/lib/libonnxruntime.dylib")
-        os.remove(f"{install_dir}/lib/libonnxruntime.so")
+
+        if is_linux():
+            os.remove(f"{install_dir}/lib/libonnxruntime.so")
+
+        if is_macos():
+            os.remove(f"{install_dir}/lib/libonnxruntime.dylib")
 
         if is_windows():
             shutil.rmtree(f"{install_dir}/lib")
